+++ pcs/utils.py
@@ -2951,7 +2951,14 @@ def get_user_and_pass():
     return username, password
 
 
-def get_input_modifiers():
+def get_input_modifiers() -> InputModifiers:
+    if "--master" in pcs_options:
+        print(pcs_options["--master"])
+        reports_output.warn(
+            "Option '--master' is deprecated and has been replaced by option "
+            "'--promoted' "
+        )
+        pcs_options["--promoted"] = pcs_options.pop("--master")
     return InputModifiers(pcs_options)