+++ pcs_test/tier0/cli/common/test_parse_args.py
@@ -604,7 +604,7 @@ class InputModifiersTest(TestCase):
             "--groups",
             "--hide-inactive",
             "--local",
-            "--master",
+            "--promoted",
             "--monitor",
             "--no-default-ops",
             "--no-expire-check",